samsung: doze: Rename to avoid conflicts with AdvancedDisplay
