+++ doze/src/org/lineageos/settings/doze/SamsungDozeActivity.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.lineageos.settings.device;
+package org.lineageos.settings.doze;
 
 import android.os.Bundle;
 import android.preference.PreferenceActivity;
