+++ doze/src/org/lineageos/settings/doze/TouchscreenGestureSettings.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.lineageos.settings.device;
+package org.lineageos.settings.doze;
 
 import android.app.ActionBar;
 import android.os.Bundle;
